+++ code/color.hpp
@@ -11,11 +11,11 @@ public:
 	Color();
 	Color(int _r, int _g, int _b);
 	Color(int value);
-	Color(float _r, float _g, float _b);
-	Color(float value);
+	Color(double _r, double _g, double _b);
+	Color(double value);
 
 	Color ToGray() const;
-	Color ToGray(float kr, float kg, float kb) const;
+	Color ToGray(double kr, double kg, double kb) const;
 
 	static Color Red()    { return Color(255, 0  , 0);   };
 	static Color Green()  { return Color(0  , 255, 0);   };
